Re-enable test
