+++ src/Components/test/E2ETest/ServerRenderingTests/FormHandlingTests/FormWithParentBindingContextTest.cs
@@ -1369,59 +1369,6 @@ public class FormWithParentBindingContextTest : ServerTestBase<BasicTestAppServe
         Assert.Equal("Total: 7", Browser.Exists(By.Id("form-collection")).Text);
     }
 
-    [Theory]
-    // [InlineData(true)] QuarantinedTest: https://github.com/dotnet/aspnetcore/issues/61882
-    [InlineData(false)]
-    public void CanUseFormWithMethodGet(bool suppressEnhancedNavigation)
-    {
-        SuppressEnhancedNavigation(suppressEnhancedNavigation);
-        GoTo("forms/method-get");
-        Browser.Equal("Form with method=get", () => Browser.FindElement(By.TagName("h2")).Text);
-
-        // Validate initial state
-        var stringInput = Browser.FindElement(By.Id("mystring"));
-        var boolInput = Browser.FindElement(By.Id("mybool"));
-        Browser.Equal("Initial value", () => stringInput.GetDomProperty("value"));
-        Browser.Equal("False", () => boolInput.GetDomProperty("checked"));
-
-        // Edit and submit the form; check it worked
-        stringInput.Clear();
-        stringInput.SendKeys("Edited value");
-        boolInput.Click();
-        Browser.FindElement(By.Id("submit-get-form")).Click();
-        AssertUiState("Edited value", true);
-        Browser.Contains($"MyString=Edited+value", () => Browser.Url);
-        Browser.Contains($"MyBool=True", () => Browser.Url);
-
-        // Check 'back' correctly gets us to the previous state
-        Browser.Navigate().Back();
-        AssertUiState("Initial value", false);
-        Browser.False(() => Browser.Url.Contains("MyString"));
-        Browser.False(() => Browser.Url.Contains("MyBool"));
-
-        // Check 'forward' correctly recreates the edited state
-        Browser.Navigate().Forward();
-        AssertUiState("Edited value", true);
-        Browser.Contains($"MyString=Edited+value", () => Browser.Url);
-        Browser.Contains($"MyBool=True", () => Browser.Url);
-
-        void AssertUiState(string expectedStringValue, bool expectedBoolValue)
-        {
-            Browser.Equal(expectedStringValue, () => Browser.FindElement(By.Id("mystring-value")).Text);
-            Browser.Equal(expectedBoolValue.ToString(), () => Browser.FindElement(By.Id("mybool-value")).Text);
-
-            // If we're not suppressing, we'll keep referencing the same elements to show they were preserved
-            if (suppressEnhancedNavigation)
-            {
-                stringInput = Browser.FindElement(By.Id("mystring"));
-                boolInput = Browser.FindElement(By.Id("mybool"));
-            }
-
-            Browser.Equal(expectedStringValue, () => stringInput.GetDomProperty("value"));
-            Browser.Equal(expectedBoolValue.ToString(), () => boolInput.GetDomProperty("checked"));
-        }
-    }
-
     [Fact]
     public void RadioButtonGetsResetAfterSubmittingEnhancedForm()
     {
